@@ -4,7 +4,8 @@ version: 0.9.1
 repository: https://github.com/gnudles/dart_i18n_generator
 
 environment:
-  sdk: '>=2.7.0 <3.0.0'
+  sdk: '>=3.0.0 <4.0.0'
+  flutter: '>=3.10.0'
 
 dependencies:
   yaml: '>=2.0.0 <4.0.0'
